Update MemberDripDropFactory.sol
Updated memberIndex of memberList to start at "1" instead of "0." Changes were made to the constructor, addMember, removeMember, and isMember. However, such implementation breaks the isMember function when a non-member is queried.  The error is "call to MemberDripDrop.isMember errored: VM error: invalid opcode."

Notice that the changes reflect manual decrements of certain values in addMember, removeMember, and isMember. These manual decrements are made in response to removing "-1" from line 143, which was originally added to accommodate the .push function (.push function starts an array at 1 instead of 0, in contrast with the mapping function, which starts an array with 0).